Switch to TypeScript on the client
