+++ client/src/Pages/AddLanguage/AddLanguage.tsx
@@ -5,14 +5,17 @@
  */
 
 import React, { useState, useEffect } from 'react';
-import { Outlet } from 'react-router-dom';
 import { Helmet, HelmetProvider } from 'react-helmet-async';
 
-const Languages = () => {
-  const [apiData, setApiData] = useState(null);
+type ApiData = {
+  title: string;
+};
+
+const AddLanguage = () => {
+  const [apiData, setApiData] = useState<ApiData | null>(null);
 
   useEffect(() => {
-    fetch('/api/languages')
+    fetch('/api/languages/add')
       .then((response) => response.json())
       .then((data) => setApiData(data))
       .catch((error) => console.error('Error fetching data:', error));
@@ -29,21 +32,21 @@ const Languages = () => {
         <title>{apiData.title}</title>
       </Helmet>
       <h1>{apiData.title}</h1>
-      <a href="/languages/add">Add language</a>
-      {apiData.languages.map((language, i) => (
-        <React.Fragment key={i}>
-          <p>{language.name}</p>
-          <a href={"/languages/edit/" + language.id}>Edit</a>
-          <form method="post" action="/api/languages/delete">
-            <input type="hidden" name="id" value={language.id} />
-            <button type="submit">Delete</button>
-          </form>
-        </React.Fragment>
-      ))}
+      <form method="post" action="/api/languages/add">
+        <label htmlFor="name">Name</label>
+        <input type="text" name="name" id="name" />
+        <br />
+        <label htmlFor="dictionaryUrl">Dictionary</label>
+        <input type="text" name="dictionaryUrl" id="dictionaryUrl" />
+        <br />
+        <label htmlFor="shouldShowSpaces">Show spaces</label>
+        <input type="checkbox" name="shouldShowSpaces" id="shouldShowSpaces" />
+        <br />
 
-      <Outlet />
+        <button type="submit">Add</button>
+      </form>
     </HelmetProvider>
   );
 };
 
-export { Languages };
+export { AddLanguage };
