+++ client/src/Pages/AddText/index.ts
