+++ client/src/Pages/Settings/Settings.tsx
@@ -7,8 +7,12 @@
 import React, { useState, useEffect } from 'react';
 import { Helmet, HelmetProvider } from 'react-helmet-async';
 
+type ApiData = {
+  title: string;
+};
+
 const Settings = () => {
-  const [apiData, setApiData] = useState(null);
+  const [apiData, setApiData] = useState<ApiData | null>(null);
 
   useEffect(() => {
     fetch('/api/settings')
