+++ client/src/Pages/Texts/Texts.tsx
@@ -5,13 +5,22 @@
  */
 
 import React, { useState, useEffect } from 'react';
+import { Outlet } from 'react-router-dom';
 import { Helmet, HelmetProvider } from 'react-helmet-async';
 
-const AddText = () => {
-  const [apiData, setApiData] = useState(null);
+type ApiData = {
+  title: string;
+  texts: {
+    id: number;
+    title: string;
+  }[];
+};
+
+const Texts = () => {
+  const [apiData, setApiData] = useState<ApiData | null>(null);
 
   useEffect(() => {
-    fetch('/api/texts/add')
+    fetch('/api/texts')
       .then((response) => response.json())
       .then((data) => setApiData(data))
       .catch((error) => console.error('Error fetching data:', error));
@@ -28,21 +37,21 @@ const AddText = () => {
         <title>{apiData.title}</title>
       </Helmet>
       <h1>{apiData.title}</h1>
-      <form method="post" action="/api/texts/add">
-        <label htmlFor="title">Title</label>
-        <input type="text" name="title" id="title" />
-        <br />
-        <label htmlFor="content">Content</label>
-        <textarea name="content" id="content"></textarea>
-        <br />
-        <label htmlFor="sourceUrl">URL</label>
-        <input type="text" name="sourceUrl" id="sourceUrl" />
-        <br />
+      <a href="/texts/add">Add text</a>
+      {apiData.texts.map((text, i) => (
+        <React.Fragment key={i}>
+          <p>{text.title}</p>
+          <a href={"/texts/edit/" + text.id}>Edit</a>
+          <form method="post" action="/api/texts/delete">
+            <input type="hidden" name="id" value={text.id} />
+            <button type="submit">Delete</button>
+          </form>
+        </React.Fragment>
+      ))}
 
-        <button type="submit">Add</button>
-      </form>
+      <Outlet />
     </HelmetProvider>
   );
 };
 
-export { AddText };
+export { Texts };
