+++ client/src/Pages/Texts/index.ts
